http://jira.aplana.com/browse/SBRFACCTAX-9285 исправил приведение типа для автонумеруемых колонок

+++ dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/impl/datarow/DataRowMapper.java
@@ -117,9 +117,9 @@ class DataRowMapper implements RowMapper<DataRow<Cell>> {
 			// Values
 			if (ColumnType.AUTO.equals(cell.getColumn().getColumnType()) && rs.getString("alias") == null) {
 				if (NumerationType.CROSS.equals(((AutoNumerationColumn) cell.getColumn()).getNumerationType())) {
-						cell.setValue(SqlUtils.getInteger(rs, "IDX2") + previousRowNumber, rowNum);
+						cell.setValue(SqlUtils.getLong(rs, "IDX2") + previousRowNumber, rowNum);
 				} else {
-					cell.setValue(SqlUtils.getInteger(rs, "IDX2"), rowNum);
+					cell.setValue(SqlUtils.getLong(rs, "IDX2"), rowNum);
 				}
 			} else {
 				DataRowDaoImplUtils.CellValueExtractor extr = getCellValueExtractor(cell.getColumn());
